# back/server/target/generated-sources/openapi/src/main/java/com/moe/myfamilybudget/api/model/VariableOverrideDto.java
package com.moe.myfamilybudget.api.model;

import java.math.BigDecimal;

public record VariableOverrideDto(
    String id,
    String label,
    Integer year,
    BigDecimal amount,
    String taxable
) {}